+++ Herdinator/src/game/actors/Dog.java
@@ -20,14 +20,14 @@ public class Dog extends MovableActor
     private static final Integer SPRITE_SHEET_SPRITE_HEIGHT = 32;
     private static final Color SPRITE_SHEET_BACKGROUND_COLOR = new Color( 123, 198, 132 );
     
-    private static final Float SPEED = 0.1f;   
+    private static final Double SPEED = 0.1;   
 
     /**
      * Constructor.
      * @param position
      * @throws SlickException 
      */
-    public Dog( Point2D.Float position ) throws SlickException
+    public Dog( Point2D.Double position ) throws SlickException
     {
         super( position, Dog.SPEED );        
     }
