+++ Herdinator/src/game/actors/Wolf.java
@@ -1,13 +1,13 @@
 package game.actors;
 
 import game.base.MovableActor;
+import game.util.SpriteSheetUtil;
 import java.awt.geom.Point2D;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import org.newdawn.slick.Color;
 import org.newdawn.slick.SlickException;
 import org.newdawn.slick.SpriteSheet;
-import game.util.SpriteSheetUtil;
 
 /**
  *
@@ -20,14 +20,14 @@ public class Wolf extends MovableActor
     private static final Integer SPRITE_SHEET_SPRITE_HEIGHT = 32;
     private static final Color SPRITE_SHEET_BACKGROUND_COLOR = new Color( 123, 198, 132 );
     
-    private static final Float SPEED = 0.1f;
+    private static final Double SPEED = 0.1;
 
     /**
      * Constructor.
      * @param position
      * @throws SlickException 
      */
-    public Wolf( Point2D.Float position ) throws SlickException
+    public Wolf( Point2D.Double position ) throws SlickException
     {
         super( position,Wolf.SPEED );
     }
@@ -62,10 +62,10 @@ public class Wolf extends MovableActor
      * @return
      */
     /*
-    private Point2D.Float locationClosestSheep(){
-        Point2D.Float locationClosestSheep = new Point2D.Float(0,0); //to make Netbeans happy
+    private Point2D.Double locationClosestSheep(){
+        Point2D.Double locationClosestSheep = new Point2D.Double(0,0); //to make Netbeans happy
         double distanceToNearestSheep = Math.pow(10, 9); //should start higher than anything
-        for(Point2D.Float sheepLocation:sheepLocations){
+        for(Point2D.Double sheepLocation:sheepLocations){
             if(euclideanDistance(sheepLocation, this.getPosition()) < distanceToNearestSheep){
                 locationClosestSheep = sheepLocation;
                 distanceToNearestSheep = euclideanDistance(sheepLocation, this.getPosition());
